compiler warning fix
git-svn-id: svn://scribus.net/branches/Version133x/Scribus@7165 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/pixmapexport/export.cpp
@@ -105,7 +105,7 @@ bool PixmapExportPlugin::run(QString target)
 		QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
 		std::vector<int> pageNs;
 		ex->pageDPI = dia->DPIBox->value();
-		ex->enlargement = dia->EnlargementBox->value();
+		ex->enlargement = static_cast<int>(dia->EnlargementBox->value());
 		ex->quality = dia->QualityBox->value();
 		ex->exportDir = dia->OutputDirectory->text();
 		ex->bitmapType = dia->bitmapType;
